+++ flexiznam/main.py
@@ -18,7 +18,7 @@ def _format_project(project_id, prm):
     return project_id
 
 
-def _lookup_project(project_id, prm):
+def lookup_project(project_id, prm):
     """
     Look up project name by hexadecimal id
     """
@@ -115,25 +115,28 @@ def add_mouse(
             mcms_username = PARAMETERS["mcms_username"]
         if mcms_animal_name is None:
             mcms_animal_name = mouse_name
-        mcms_info = dict(
-            mcms.get_mouse_df(
-                mouse_name=mcms_animal_name,
-                username=mcms_username,
-                password=mcms_password,
-            )
+        mcms_info = mcms.get_mouse_info(
+            mouse_name=mcms_animal_name,
+            username=mcms_username,
+            password=mcms_password,
         )
-        # format properly results
-        for k, v in mcms_info.items():
-            if type(v) != str:
-                mcms_info[k] = float(v)
-            else:
-                mcms_info[k] = v.strip()
-
+        # flatten alleles and colony
+        alleles = mcms_info.pop("alleles")
+        for gene in alleles:
+            gene_name = gene["allele"]["shortAlleleSymbol"].replace(" ", "_")
+            mcms_info[gene_name] = gene["genotype"]["name"]
+        colony = mcms_info.pop("colony")
+        mcms_info["colony"] = colony["name"]
+        mcms_info["colony_prefix"] = colony["colonyPrefix"]
+        if not mcms_info:
+            raise IOError(f"Could not get info for mouse {mouse_name} from MCMS")
         # update mouse_info with mcms_info but prioritise mouse_info for conflicts
         mouse_info = dict(mcms_info, **mouse_info)
 
     # add the genealogy info, which is just [mouse_name]
     mouse_info["genealogy"] = [mouse_name]
+    project_name = lookup_project(flexilims_session.project_id, PARAMETERS)
+    mouse_info["path"] = str(Path(project_name) / mouse_name)
     resp = flexilims_session.post(
         datatype="mouse",
         name=mouse_name,
@@ -969,7 +972,7 @@ def get_datasets(
     if flexilims_session is None:
         flexilims_session = get_flexilims_session(project_id)
     else:
-        project_id = _lookup_project(flexilims_session.project_id, PARAMETERS)
+        project_id = lookup_project(flexilims_session.project_id, PARAMETERS)
     recordings = get_entities(
         datatype="recording",
         origin_id=origin_id,
@@ -989,7 +992,7 @@ def get_datasets(
             flexilims_session=flexilims_session,
         )
         datapaths = []
-        for (dataset_path, is_raw) in zip(datasets["path"], datasets["is_raw"]):
+        for dataset_path, is_raw in zip(datasets["path"], datasets["is_raw"]):
             prefix = (
                 PARAMETERS["data_root"]["raw"]
                 if is_raw == "yes"