+++ flexiznam/schema/camera_data.py
@@ -137,6 +137,7 @@ class CameraData(Dataset):
         project=None,
         project_id=None,
         origin_id=None,
+        id=None,
         flexilims_session=None,
     ):
         """Create a camera dataset
@@ -154,6 +155,7 @@ class CameraData(Dataset):
             project_id: hexadecimal code for the project. Must be in config, can be
                         guessed from project
             origin_id: hexadecimal code for the origin on flexilims.
+            id: hexadecimal code for the dataset on flexilims.
             flexilims_session: authentication session to connect to flexilims
 
 
@@ -180,6 +182,7 @@ class CameraData(Dataset):
             project=project,
             project_id=project_id,
             origin_id=origin_id,
+            id=id,
             flexilims_session=flexilims_session,
         )